Refactor anitya.app to remove Flask routes
This refactors the ``anitya.app`` module to remove the Flask route
handlers. These have been moved to the ``anitya.ui`` module.
``anitya.app`` has been documented to be exclusively for application
setup.
